Remove Generators from Factory
